requirements.txt
python-telegram-bot
faiss-cpu
numpy
agno
sentence_transformers